SOCINT-56 introduce other feature files. Some renaming of packages/files to fit with cucumber accepted terminology
