+++ pom.xml
@@ -17,7 +17,6 @@
 
   <properties>
     <cucumber.version>6.10.4</cucumber.version>
-    <!-- <cucumber.version>5.4.0</cucumber.version> -->
   </properties>
 
   <scm>
@@ -148,7 +147,7 @@
         <configuration>
           <testSourceDirectory>${project.basedir}/src/test/java</testSourceDirectory>
           <includes>
-            <include>**/glue/*.java</include>
+            <include>**/run/*.java</include>
           </includes>
         </configuration>
       </plugin>
