+++ src/test/java/uk/gov/ons/ctp/integration/rhcucumber/glue/RateLimiterMock.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.integration.rhcucumber.steps;
+package uk.gov.ons.ctp.integration.rhcucumber.glue;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
