--- src/test/java/uk/gov/ons/ctp/integration/rhcucumber/glue/RequestUAC.java
@@ -1,14 +1,0 @@
-package uk.gov.ons.ctp.integration.rhcucumber.glue;
-
-import io.cucumber.junit.Cucumber;
-import io.cucumber.junit.CucumberOptions;
-import org.junit.runner.RunWith;
-import uk.gov.ons.ctp.integration.rhcucumber.GlueConst;
-
-@RunWith(Cucumber.class)
-@CucumberOptions(
-    plugin = {"pretty", "html:target/cucumber/request-uac"},
-    features = {"src/test/resources/features/Request-UAC.feature"},
-    tags = GlueConst.COMMON_TAGS,
-    glue = {"uk.gov.ons.ctp.integration.rhcucumber.steps"})
-public class RequestUAC {}
